backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
passlib==1.7.4
python-jose==3.3.0
argon2-cffi==23.1.0
requests==2.31.0
redis==5.0.1